Update transformer.ipynb
Hey! I really liked your implementation here, and took some inspiration for some of my own projects! The only thing bugging me is the scale factor in the MultiHeadAttention. I believe it's more optimal to divide by the size of each head, instead of the whole self.d_model. Also, if you divided by this factor in the forward(), I think you should also divide by the factor in the backprop(), not multiply by it. That's because the partial derivative through (Const * f(x)) is (Const * d(f(x))/dx). But if you did that for a reason, I would love to hear why that is! I've started to learn transformers only relatively recently. But I think that's it - just giving a heads up, because these changes might make the gradient flow a lot less spiky if it does end up working!